tested and fixed version of Quest2
actually we do not allow newer versions than the newest stable due to
beta versions or inofficial forks are in general not a good idea
(especially for security-relevant tasks). A fork would require it's own
audit (which includes an audit of the code of that fork!).

considering #11 as solved by now...